Fix: generation/tick should be uint64_t instead of size_t

+++ core/index/segment_writer.hpp
@@ -71,7 +71,7 @@ class segment_writer : util::noncopyable {
 
  public:
   struct update_context {
-    size_t generation;
+    uint64_t generation;
     size_t update_id;
   };
